requirements.txt
langchain>=0.1.0
python-dotenv>=0.19.0
requests>=2.25.1
flask>=2.0.1
click>=8.0.1
pytest>=6.2.4
langchain-openai